Removed the external session ID field.
Several systems inside np1sec relied on messages having a per-room-unique ID, consistent between the different participants in the room, to be supplied by the carrier chat system.
This doesn't actually exist for any currently implemented chat systems and wasn't implemented, causing the systems to rely on it to break.
I have replaced it with a weaker per-session message ID; I hope that doesn't break anything.