Implemented channel event queue.
Channels now contain a queue of expected messages. Whenever a message of a
certain type arrives, it is matched against the head of the message queue;
if this is not the expected message, or it does not contain the expected
content, this will cause the sender to get kicked from the channel.

Currently the only message of this form is the channel status message. Later
messages that fit the same scheme include the different messages involved in
the key negotiation process, and possibly others.